Basic Functionality Ready
This has basic functionality set up and can be run successfully.

NOTE: The first run of this wound up with LOTS of events. This is likely due to the fact that the spreadsheet was meant to be loaded with parameters at the "family" level. Added to the extract_events function to divide the probabilities by the number of family members.

It needs a little debugging. Specifically:

-Need to figure out disparity between max values reported. I believe this is failing to include one month's premium.
-Need to add some cells to the spreadsheet to calculate expected value for each plan.
-Continued debug of results and comparison to old simulation.

Future Work:
-Revise spreadsheet so each column on the sheet has a different family member and THEIR probabilities.
-Add functions for DETERMINISTIC (vice stochastic) events.